pyspark_project_template/requirements.txt
pyspark
pyyaml
pytest
chispa